third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coep.https-expected.txt

This is a testharness.js-based test.
PASS Same-origin popup with coop/coep
FAIL historical: "same-site" popup with coop/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Same-origin popup with coop unsafe-none/coep assert_equals: Openee browsing context closed expected true but got false
FAIL historical: "same-site" popup with coop unsafe-none/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Same-origin popup with coop unsafe-none without coep assert_equals: Openee browsing context closed expected true but got false
FAIL historical: "same-site" popup with coop unsafe-none without coep assert_equals: Openee browsing context closed expected true but got false
FAIL Same-origin popup without coep assert_equals: Openee browsing context closed expected true but got false
FAIL historical: "same-site" popup without coep assert_equals: Openee browsing context closed expected true but got false
FAIL Bonus: window.crossOriginIsolated assert_true: expected true got undefined
Harness: the test ran to completion.
